Renames and formats
