+++ zio-http/src/main/scala/zio/http/api/HeaderCodec.scala
@@ -2,4 +2,4 @@ package zio.http.api
 
 import zio.stacktracer.TracingImplicits.disableAutoTrace // scalafix:ok;
 
-object RouteCodec extends RouteCodecs
+object HeaderCodec extends HeaderCodecs
