+++ zio-http/src/main/scala/zio/http/api/QueryCodec.scala
@@ -2,16 +2,4 @@ package zio.http.api
 
 import zio.stacktracer.TracingImplicits.disableAutoTrace // scalafix:ok;
 
-object QueryCodec {
-  def query(name: String): QueryCodec[String] =
-    HttpCodec.Query(name, TextCodec.string)
-
-  def queryBool(name: String): QueryCodec[Boolean] =
-    HttpCodec.Query(name, TextCodec.boolean)
-
-  def queryInt(name: String): QueryCodec[Int] =
-    HttpCodec.Query(name, TextCodec.int)
-
-  def queryAs[A](name: String)(implicit codec: TextCodec[A]): QueryCodec[A] =
-    HttpCodec.Query(name, codec)
-}
+object QueryCodec extends QueryCodecs
