+++ zio-http/src/main/scala/zio/http/api/RouteCodecs.scala
@@ -1,0 +1,18 @@
+package zio.http.api
+
+import java.util.UUID
+import zio.stacktracer.TracingImplicits.disableAutoTrace // scalafix:ok;
+
+private[api] trait RouteCodecs {
+  def literal(string: String): RouteCodec[Unit] =
+    HttpCodec.Route(TextCodec.constant(string))
+
+  val int: RouteCodec[Int] =
+    HttpCodec.Route(TextCodec.int)
+
+  val string: RouteCodec[String] =
+    HttpCodec.Route(TextCodec.string)
+
+  val uuid: RouteCodec[UUID] =
+    HttpCodec.Route(TextCodec.uuid)
+}
